Move the ImageNavigatorFragment to its own package
